fix: use size_t instead

+++ lab/cnn/src/layer/custom/new-forward.cc
@@ -98,8 +98,8 @@ void GPUInterface::conv_forward_gpu(cl_mem device_y, const cl_mem device_x, cons
     int H_out = H - K + 1;
     int W_out = W - K + 1;
 
-    int W_grid = (W_out + TILE_WIDTH - 1) / TILE_WIDTH;
-    int H_grid = (H_out + TILE_WIDTH - 1) / TILE_WIDTH;
+    size_t W_grid = (W_out + TILE_WIDTH - 1) / TILE_WIDTH;
+    size_t H_grid = (H_out + TILE_WIDTH - 1) / TILE_WIDTH;
 
     size_t global_item_size[3] = {M * TILE_WIDTH, W_grid * H_grid * TILE_WIDTH, B}; 
     size_t local_item_size[3] = {TILE_WIDTH, TILE_WIDTH, 1};
